part4 complete

+++ part4/controllers/blogs.js
@@ -1,8 +1,11 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
-const Blog = require('../models/blog.js')
+const Blog = require('../models/blog')
+// const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('author', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -14,15 +17,36 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   if (!request.body.title || !request.body.url) {
     return response.status(400).end()
-  } // else {
-  const blog = new Blog(request.body)
+  } 
+  const body = request.body
 
-  const result = await blog.save()
-  response.status(201).json(result)
-  // }
+  if(!request.user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+
+  const blog = new Blog({
+    title: body.title,
+    author: request.user,
+    url: body.url,
+    likes: body.likes
+  })
+
+  const savedBlog = await blog.save()
+
+  request.user.blogs = request.user.blogs.concat(savedBlog._id)
+  await request.user.save()
+
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  if (!request.user) {
+    return response.status(401).json({error: 'authentication token must be provided to delete a blog'})
+  }
+  if (!(request.user.blogs.includes(request.params.id))) {
+    return response.status(401).json({ error: 'invalid user: you can only delete blogs uploaded by yourself' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
